Refactor project structure: reorganized into layouts, processing, callbacks; split DataProcessor into FileProcessor, TableProcessor, MapProcessor
