+++ components.py
@@ -1,0 +1,2 @@
+# Пустой файл для будущих общих компонентов
+# Например, константы для таблиц или конфигурации
