+++ utils.py
@@ -1,0 +1,2 @@
+# Пустой файл для утилитарных функций
+# Будет заполнен при необходимости
